feat(telemetry): process-wide pipeline, per-Room sessions, three instruments
Telemetry now starts with the SDK, not with connect(): `LiveKitSDK.setTelemetry(options)`
configures a process-wide `TelemetryHub` (core pipeline, transport, destination, log
relay sink), so pre-connect work - track creation, permission failures, connect
attempts that never reach a server - is captured. `RoomOptions.telemetry` is gone.

Each Room gets a session (own trace id) from the hub in `init`, exposed as
`room.telemetryTraceId`; `emitTelemetryEvent` works before connect. The
destination is derived from the first connect (`https://<host>/observability/logs/otlp/v0`,
`Authorization: Bearer <token>`) unless the options name an endpoint; until then
everything is buffered on device.

The instruments are three independent components, one per area of the design doc:
`DeviceTelemetry` (process-level: thermal, low power, memory, network, battery,
app state, audio session), `RoomTelemetry` (session identity, spans, subscribe spans,
app-defined events) and `RTCTelemetry` (track statistics -> session; one signal out,
`onFirstMedia`, which the Room wires to the subscribe spans). `ConnectionDependencies`
only settles open spans and flushes when a connection ends.
