+++ Tests/LiveKitTestSupport/TestEnvironment.swift
@@ -94,7 +94,7 @@ public enum TestEnvironment {
             let encryptionOptions = $0.element.isE2eeEnabled
                 ? $0.element.encryptionOptions ?? EncryptionOptions(keyProvider: BaseKeyProvider(isSharedKey: true, sharedKey: sharedKey))
                 : nil
-            let roomOptions = RoomOptions(encryptionOptions: encryptionOptions, reportRemoteTrackStatistics: true, singlePeerConnection: $0.element.singlePeerConnection, telemetry: $0.element.telemetry)
+            let roomOptions = RoomOptions(encryptionOptions: encryptionOptions, reportRemoteTrackStatistics: true, singlePeerConnection: $0.element.singlePeerConnection)
 
             let room = Room(delegate: $0.element.delegate, connectOptions: connectOptions, roomOptions: roomOptions)
             let roomName = $0.element.roomName ?? sharedRoomName
